refactor(linker): isolate apply and clean orchestration
